+++ src/lib.rs
@@ -24,7 +24,7 @@
 //!
 //! impl ConsumerDelegate for Subscriber {
 //!   fn on_new_delivery(&self, delivery: Delivery) {
-//!     self.channel.basic_ack(delivery.delivery_tag, BasicAckOptions::default()).into_error().expect("basic_ack");
+//!     self.channel.basic_ack(delivery.delivery_tag, BasicAckOptions::default()).wait().expect("basic_ack");
 //!   }
 //! }
 //!